feat: reemplazar el diálogo de información general del presupuesto y eliminar registros de consola
