+++ src/hooks/use-subworkitem.ts
@@ -37,7 +37,6 @@ export const useSubWorkItem = (options: UseSubWorkItemsProps = {}) => {
             new Promise(async (resolve, reject) => {
                 try {
                     const result = await create(input);
-                    console.log(result);
                     if (
                         result.error &&
                         typeof result.error === "object" &&
